+++ src/components/admin-page/NavBar/NavBar.module.scss
@@ -101,7 +101,7 @@
     margin: 5px;
     display: flex;
     align-items: center;
-    gap: 5px;
+    gap: 10px;
     padding: 5px 12px;
     position: relative;
     &::before {
@@ -129,10 +129,9 @@
 
     .icon{
       display: block;
-      width: 25px;
-      height: 25px;
-      //stroke:#fff;
-      //fill:#fff;
+      width: 20px;
+      height: 20px;
+      margin-bottom: 2px;
     }
     .text._hide{
       display: none;
