+++ src/components/login-page/LoginSection/LoginSection.jsx
@@ -8,6 +8,7 @@ import Loader from '../../shared/loader/Loader';
 import stateUseAlert from '@/src/state/stateUseAlert';
 import UseAlert from '../../shared/UseAlert/UseAlert';
 import LoginForm from './loginForm/loginForm';
+import Section from '../Section/Section';
 
 export default function LoginSection() {
   const open = stateUseAlert(state => state.open);
@@ -37,27 +38,19 @@ export default function LoginSection() {
   useEffect(() => {
     const token = sessionStorage.getItem('access_token');
     if (token) {
-      router.replace('/admin');
+      //router.replace('/admin');
     }
   });
 
   return (
-    <section className={styles.section}>
-      <div className={styles.wrapper}>
-        
-        <div className={styles.title}>
-          <h2>Вхід</h2>
-          <p>Введіть дані для входу</p>
-        </div>
-        
-        <LoginForm handleMutate={mutate} isSuccess={data?.token}/>
+    <Section title={'Вхід'} text={ 'Введіть дані для входу' }>
+      <LoginForm handleMutate={mutate} isSuccess={data?.token}/>
 
-        <Link href={'/admin/forgot-password'}>Забули пароль?</Link>
+      <Link className={styles.link} href={'/admin/forgot-password'}>Забули пароль?</Link>
 
-        {isPending && <Loader/>} 
+      {isPending && <Loader/>} 
 
-        <UseAlert title='Невірний логін або пароль' text='Перевірте данні та спробуйте ще'/>
-      </div>
-    </section>
+      <UseAlert title='Невірний логін або пароль' text='Перевірте данні та спробуйте ще'/>
+    </Section>
   )
 }
