+++ src/components/login-page/LoginSection/LoginSection.module.scss
@@ -1,58 +1,17 @@
 @use "../../../styles/mixins" as mix;
 @use "../../../styles/variables" as var;
 
-.section{
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  min-height: 100vh;
-  padding: 10px;
+.link{
+  @include mix.apply-font-styles(var.$h2-font);
+  font-weight: 400;
+  font-size: 20px;
+  background: var.$icons-gradient;
+  background-clip: text;
+    -webkit-background-clip: text;
+    -webkit-text-fill-color: transparent;
+  transition: all var.$transition-duration-base var.$transition-timing-base;
+}
 
-  .wrapper{
-    border-radius: 8px;
-    background: #d2d2d2;
-    color: #020202;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    padding: 30px 50px;
-    gap: 36px;
-    @media (min-height: 800px) {
-      padding: 50px;
-      gap: 36px;
-    }
-    .title{
-      text-align: center;
-      & > h2{
-        @include mix.apply-font-styles(var.$h1-font);
-        font-weight: 600;
-        font-size: 40px;
-      }
-  
-      & > p{
-        @include mix.apply-font-styles(var.$h2-font);
-        font-weight: 400;
-        font-size: 20px;
-      }
-    }
-
-    .link{
-      @include mix.apply-font-styles(var.$h2-font);
-      font-weight: 400;
-      font-size: 20px;
-    }
-    &  a {
-      @include mix.apply-font-styles(var.$h2-font);
-      font-weight: 400;
-      font-size: 20px;
-      background: var.$icons-gradient;
-      background-clip: text;
-        -webkit-background-clip: text;
-        -webkit-text-fill-color: transparent;
-      transition: all .3s;
-    }
-    &  a:hover{
-      opacity: .7;
-    }
-  }
+.link:hover{
+  opacity: .7;
 }
