+++ src/components/login-page/Registration/Registration.jsx
@@ -1,19 +1,14 @@
 "use client";
-
-import { Link, useRouter } from '@/src/navigation';
-import InputField from '../../shared/InputField/InputField';
-import Loader from '../../shared/loader/Loader';
-import MainButton from '../../shared/MainButton/MainButton';
 import styles from './Registration.module.scss';
-import { Icon } from '../../shared/Icon/Icon';
-import { registrationDefaultValues, registrationSchema } from './registrationScheme';
-import { useForm } from "react-hook-form";
-import { zodResolver } from "@hookform/resolvers/zod";
-import { useEffect, useState } from 'react';
-import stateUseAlert from '@/src/state/stateUseAlert';
-import { useMutation } from '@tanstack/react-query';
+import { useEffect } from 'react';
 import { registerAdmin } from '@/src/api/auth';
+import { Link, useRouter } from '@/src/navigation';
+import { useMutation } from '@tanstack/react-query';
+import stateUseAlert from '@/src/state/stateUseAlert';
+import Loader from '../../shared/loader/Loader';
 import UseAlert from '../../shared/UseAlert/UseAlert';
+import Section from '../Section/Section';
+import RegistrationForm from './RegistrationForm/RegistrationForm';
 
 export default function Registration() {
   const open = stateUseAlert(state => state.open);
@@ -25,41 +20,11 @@ export default function Registration() {
     },
   })
 
-  // if(data){
-  //   //router.replace('/login/signIn');
-  //   open('success')
-  //   console.log(data)
-  // }
-
-
-
-  // if(isError){
-  //   //router.replace('/login/signIn');
-  
-  //   //open('error')
-  // }
-
-
-  const {
-    register,
-    handleSubmit,
-    formState: { errors, isValid },
-    reset
-  } = useForm({ defaultValues: {...registrationDefaultValues}, resolver: zodResolver(registrationSchema), mode: 'onBlur'});
-
-  const [ visible, setVisible ] = useState(false);
-  const [ visible1, setVisible1 ] = useState(false);
-
-  const resetForm = () => {
-    setVisible(false)
-    setVisible1(false)
-    reset();
-  }
   useEffect(() => {
     if (data) {
-      router.replace('/admin/login')
-      resetForm()
+      //router.replace('/admin/login')
       open('success')
+      console.log(data)
     }
     if (isError) {
       console.log( error?.message)
@@ -70,103 +35,18 @@ export default function Registration() {
   },[data,isError]);
 
   const onSubmit = (data) => {
-    mutate({email:data.email, password:data.password, name:'admin' })
-    //resetForm()
-  };
-
-
-  // {
-  //   "email": "user@example.com",
-  //   "password": "password123",
-  //   "name": "John"
-  // }
-  const isDisabled = () => {
-    if (Object.keys(errors).length > 0) {
-      return true;
-    } else if (!isValid) {
-      return true;
-    } else return false;
+    mutate({email:data.email, password:data.password, name:data.email })
   };
 
   return (
-  <section className={styles.section}>
-    <form onSubmit={handleSubmit(onSubmit)} className={styles.form}>
-      <div className={styles.title}>
-        <h2>Реєстрація</h2>
-        <p>Зареєструйтесь в системі</p>
-      </div>
-      <ul className={styles.list}>
-        <li>
-          <InputField
-            id={"email"}
-            maxLength={55}
-            className={styles.item}
-            required={false}
-            //type='email'
-            placeholder={"Електронна пошта"}
-            registerOptions={register("email", { ...registrationSchema.email })}
-            isError={errors.email}
-            isValid={isValid}
-            version={"input"}
-            label={'Електронна пошта'}
-          />
-          {errors.email && <p className={styles.error_modal}>{errors.email.message}</p>}
-        </li>
-        <li className={styles.list_item} >
-          <InputField
-            id={"password"}
-            maxLength={55}
-            required={false}
-            className={styles.item}
-            type={visible?'text':'password'}
-            placeholder={"Пароль"}
-            registerOptions={register("password", { ...registrationSchema.password })}
-            isError={errors.password}
-            isValid={isValid}
-            version={"input"}
-            label={'Пароль'}
-          />
-          <button type='button' className={styles.btn} onClick={()=>{setVisible(!visible)}}>
-            <Icon width={24} height={24} name={visible?'open_eye':'closed_eye'}/>
-          </button>
-          {errors.password && <p className={styles.error_modal}>{errors.password.message}</p>}
-        </li>
-        <li className={styles.list_item}>
-          <InputField
-            id={"confirm_password"}
-            required={false}
-            maxLength={55}
-            className={styles.item}
-            type={visible1?'text':'password'}
-            placeholder={"Пароль"}
-            registerOptions={register("confirmPassword", { ...registrationSchema.confirmPassword })}
-            isError={errors.confirmPassword?.message}
-            isValid={isValid}
-            version={"input"}
-            label={'Підтвердіть пароль'}
-          />
-          <button type='button' className={styles.btn} onClick={()=>{setVisible1(!visible1)}}>
-            <Icon width={24} height={24} name={visible1?'open_eye':'closed_eye'}/>
-          </button>
-         
-          {errors.confirmPassword && <p className={styles.error_modal}>{errors.confirmPassword.message}</p>}
-        </li>
+    <Section title={'Реєстрація'} text={'Зареєструйтесь в системі'}>
+      <RegistrationForm onSubmit={onSubmit} isSuccess={data?.token}/>
 
+      <p className={styles.link}>Ви маєте акаунт? <Link href={'/admin/login'}>Авторизуватись</Link></p>
 
-      </ul>
-      <MainButton
-        type="submit"
-        disabled={isDisabled()}
-        className={styles.submit}
-      >
-        {'Зареєструватись'}
-      </MainButton>
-
-      <p>Ви маєте акаунт? <Link href={'/admin/login'}>Авторизуватись</Link></p>
       <UseAlert/>
-      {/* {isError||data && <UseAlert/>} */}
+
       {isPending && <Loader/>} 
-    </form>
-  </section>
+    </Section>
   )
 }
