+++ src/components/login-page/Registration/Registration.module.scss
@@ -1,110 +1,21 @@
 @use "../../../styles/mixins" as mix;
 @use "../../../styles/variables" as var;
 
-.section{
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  min-height: 100vh;
- 
-  padding: 10px;
-  @media (min-height: 1024px) {
-    padding: 50px 16px;
+.link{
+  @include mix.apply-font-styles(var.$h2-font);
+  font-weight: 400;
+  font-size: 20px;
+
+  & > a {
+    font-weight: 400;
+    font-size: 20px;
+    background: var.$icons-gradient;
+    background-clip: text;
+     -webkit-background-clip: text;
+     -webkit-text-fill-color: transparent;
+    transition: all var.$transition-duration-base var.$transition-timing-base;
   }
-
-  .form{
-    border-radius: 8px;
-    background: #d2d2d2;
-    color: #020202;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 18px;
-    padding: 20px 50px;
-
-    @media (min-height: 1024px) {
-      padding: 50px;
-      gap: 36px;
-    }
-
-    .title{
-      text-align: center;
-      & > h2{
-        @include mix.apply-font-styles(var.$h1-font);
-        font-weight: 600;
-        font-size: 34px;
-
-        @media (min-height: 1024px) {
-          font-size: 40px;
-        }
-      }
-  
-      & > p{
-        @include mix.apply-font-styles(var.$h2-font);
-        font-weight: 400;
-        font-size: 20px;
-      }
-    }
-    & > p{
-      @include mix.apply-font-styles(var.$h2-font);
-      font-weight: 400;
-      font-size: 20px;
-
-      & > a {
-        font-weight: 400;
-        font-size: 20px;
-        background: var.$icons-gradient;
-        background-clip: text;
-         -webkit-background-clip: text;
-         -webkit-text-fill-color: transparent;
-        transition: all .3s;
-      }
-      & > a:hover{
-        opacity: .7;
-      }
-
-    }
-    .list {
-      display: flex;
-      flex-direction: column;
-      //gap: 20px;
-      gap: 14px;
-
-      @media (min-height: 1024px)  {
-        // height: clamp(550px, 86vh, 860px);
-        gap: 20px;
-      }
-
-      .list_item{
-        position: relative;
-        .btn{
-          position: absolute;
-          right: 10px;
-          top: 40px;
-        }
-      }
-
-      .item {
-        width: clamp(280px, 84vw, 360px);
-        display: flex;
-        flex-direction: column;
-        position: relative;
-        & > input{
-          padding: 10px 35px 10px 10px;
-        }
-        & > label{
-          font-weight: 600;
-          font-size: 18px;
-        }
-
-
-      }
-
-      .error_modal{
-        @include mix.apply-font-styles(var.$text-16-font);
-        font-size: 14px;
-        color: var.$secondary-error;
-      }
-    }
+  & > a:hover{
+    opacity: .7;
   }
 }
