+++ src/components/login-page/Section/Section.module.scss
@@ -1,0 +1,40 @@
+@use "../../../styles/mixins" as mix;
+@use "../../../styles/variables" as var;
+
+.section{
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  min-height: 100vh;
+  padding: 10px;
+
+  .wrapper{
+    border-radius: 8px;
+    background: #d2d2d2;
+    color: #020202;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    padding: 20px 50px;
+    gap: 20px;
+
+    @media (min-height: 800px) {
+      padding: 50px;
+      gap: 36px;
+    }
+    .title{
+      text-align: center;
+      & > h2{
+        @include mix.apply-font-styles(var.$h1-font);
+        font-weight: 600;
+        font-size: 40px;
+      }
+  
+      & > p{
+        @include mix.apply-font-styles(var.$h2-font);
+        font-weight: 400;
+        font-size: 20px;
+      }
+    }
+  }
+}
